requirements.txt
# Core Django
Django==4.2.7
psycopg2-binary==2.9.9

# Environment variables
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5

# CORS headers
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.7

# Image handling
Pillow==10.1.0

# Additional utilities
python-decouple==3.8
